// src/app/router/types.ts
export enum AppRoutes {
  HOME = 'home',
  CATEGORIES = 'CATEGORIES',
  CATEGORY_CREATE = 'CATEGORY-CREATE',
  CATEGORY_DETAILS = 'CATEGORY-DETAILS',
  BRANDS = 'BRANDS',
  BRAND_CREATE = 'BRAND_CREATE',
  BRAND_DETAILS = 'BRAND_DETAILS',
  ATTRIBUTES = 'ATTRIBUTES',
  ATTRIBUTE_CREATE = 'ATTRIBUTE_CREATE',
  ATTRIBUTE_DETAILS = 'ATTRIBUTE_DETAILS',
}

export enum AppLayouts {
  APP = 'APP',
}